requirements.txt
# Core Dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
openai>=1.12.0

# Vector Database & Search
chromadb>=0.4.20
sentence-transformers>=2.2.2

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Database & Auth (MVP - Local Storage)
# supabase>=2.0.0  # Removed for MVP
# psycopg2-binary>=2.9.5  # Removed for MVP

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Monitoring & Logging
langsmith>=0.1.0
sentry-sdk>=1.40.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0